HEEDLS-505 Move DelegateTableRow and BulkUploadResult to separate files
