+++ DigitalLearningSolutions.Web/ViewModels/TrackingSystem/Delegates/BulkUploadResultsViewModel.cs
@@ -3,7 +3,7 @@
     using System;
     using System.Collections.Generic;
     using System.Linq;
-    using DigitalLearningSolutions.Data.Services;
+    using DigitalLearningSolutions.Data.Models.DelegateUpload;
 
     public class BulkUploadResultsViewModel
     {
